Added Archive/ExtractWithContext() and Written() to cancel and check progress
Written() returns the bytes and entries written to either disk (for an
Extractor) or an archive file (for an Archiver). This allows callers to get
progress updates.

ArchiveWithContext() and ExtractWithContext() allow users to set deadlines or
cancel archiving or extraction processes. In combination with Written() these
functions can be used to prevent resource exhaustion (zip bombs etc).