Rate-limit active go routines
The filepool's Get()/Put() was designed to rate limit up to a certain
concurrency, and whilst the unit of work is rate limited, we launch as many go
routines as there are files to be archived first. By moving the Get() call
outside of a go routine, we also limit the amount of go routines created at once.

This likely has memory savings, but also allows the "-race" detector to work
when archiving many thousands of files.